Delete extraneous. Small refactor.
