--- lib/.gitignore
@@ -1,1 +1,0 @@
-massager_parameters_ex
